符号性修复落地(审计REAL_BUG 6/7): A24F根因int8_t(治愈b1两处不可达分支+ls:18236重入保护)+b3r:7129改int16_t读+thread:382 INFINITE哨兵(int)a7+b7c:5471读侧int16_t符号扩展; build/link绿

+++ core_thread.c
@@ -379,7 +379,7 @@ uint64_t PECMD_StartWorkerThread(void *script, void **pref, uint32_t a3, uint64_
         LeaveCriticalSection(&g_csInit);
         ResumeThread(hThread);
     }
-    if ((int64_t)a7 != -1) {
+    if ((int)a7 != -1) {
         a7 = a7 + 5 + a7 / 10;
     }
     if (hThread) {
